refactor: rename components to align with figma designs
