--- src/features/view-nft/DetailsCard/index.ts
@@ -1,1 +1,0 @@
-export * from './DetailsCard';
