+++ src/features/view-nft/NFTBannerContainer/NFTBannerContainer.tsx
@@ -6,13 +6,13 @@ import { useDomainMetadata } from '../../../lib/hooks/useDomainMetadata';
 
 import { IpfsMedia } from '@zero-tech/zapp-utils/components';
 
-import styles from './Banner.module.scss';
+import styles from './NFTBannerContainer.module.scss';
 
 export interface BannerProps {
 	zna: string;
 }
 
-export const Banner: FC<BannerProps> = ({ zna }) => {
+export const NFTBannerContainer: FC<BannerProps> = ({ zna }) => {
 	const domainId = getDomainId(zna);
 
 	const { data: metadata } = useDomainMetadata(domainId);
