+++ src/features/view-nft/NFTDetailsCard/NFTDetailsCard.module.scss
