+++ src/features/view-nft/NFTDetailsCard/index.ts
@@ -1,0 +1,1 @@
+export * from './NFTDetailsCard';
