+++ src/features/view-nft/NFTDetailsCard/ui/CTAContainer/CTAContainer.module.scss
