+++ src/features/view-nft/NFTDetailsCard/ui/CTAContainer/CTAContainer.tsx
