+++ src/features/view-nft/NFTDetailsCard/ui/CTAContainer/index.ts
