+++ src/features/view-nft/NFTDetailsCard/ui/Options/Options.module.scss
