+++ src/features/view-nft/NFTDetailsCard/ui/Options/index.ts
