+++ src/features/view-nft/NFTDetailsCard/ui/index.ts
