+++ src/features/view-nft/index.ts
@@ -1,8 +1,8 @@
 export * from './Action';
 export * from './Actions';
-export * from './Banner';
-export * from './DetailsCard';
 export * from './HistoryItem';
 export * from './HistoryList';
+export * from './NFTBannerContainer';
+export * from './NFTDetailsCard';
 export * from './NFTMetrics';
 export * from './TokenHashInfo';
