--- src/features/view-subdomains/DetailsCard/index.ts
@@ -1,1 +1,0 @@
-export * from './DetailsCard';
