+++ src/features/view-subdomains/DomainBannerContainer/DomainBannerContainer.tsx
@@ -5,13 +5,13 @@ import { useDomainMetadata } from '../../../lib/hooks/useDomainMetadata';
 
 import { IpfsMedia } from '@zero-tech/zapp-utils/components';
 
-import styles from './Banner.module.scss';
+import styles from './DomainBannerContainer.module.scss';
 
 export interface BannerProps {
 	zna: string;
 }
 
-export const Banner: FC<BannerProps> = ({ zna }) => {
+export const DomainBannerContainer: FC<BannerProps> = ({ zna }) => {
 	const domainId = getDomainId(zna);
 
 	const { data: metadata } = useDomainMetadata(domainId);
