+++ src/features/view-subdomains/DomainBannerContainer/index.ts
@@ -1,0 +1,1 @@
+export * from './DomainBannerContainer';
