+++ src/features/view-subdomains/DomainDetailsCard/DomainDetailsCard.module.scss
