+++ src/features/view-subdomains/index.ts
@@ -1,5 +1,5 @@
-export * from './Banner';
-export * from './DetailsCard';
+export * from './DomainBannerContainer';
+export * from './DomainDetailsCard';
 export * from './SubdomainMetrics';
 export * from './SubdomainTable';
 export * from './SubdomainTableRow';
